requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-oauth-toolkit==1.7.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
requests==2.31.0
python-decouple==3.8
coverage==7.3.2
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
black==23.10.1
flake8==6.1.0
gunicorn==21.2.0
africastalking==1.2.5
pyjwt==2.8.0
cryptography==41.0.7
django-extensions==3.2.3
dj-database-url==2.1.0
whitenoise==6.6.0
